refactors related_data and related_data_anonymized to use related_data_helper
