+++ classes/dataset_anonymized.php
@@ -31,17 +31,6 @@ require_once(__DIR__ . '/idmap.php');
  * Class for the anonymized complete dataset evidence item.
  */
 class dataset_anonymized extends dataset {
-    /** Create idmap from a dataset of a specific type.
-     *
-     * @param array $dataset
-     * @return idmap
-     * @throws Exception
-     * @throws Exception
-     */
-    public static function create_idmap(array $dataset) : idmap {
-        $originalids = dataset_helper::get_ids_used_in_dataset($dataset);
-        return idmap::create_from_ids($originalids);
-    }
 
     /**
      * Retrieve all available analysable samples, calculate features and label.
