+++ classes/output/evidence_item.php
@@ -29,6 +29,7 @@ use templatable;
 use renderable;
 use stdClass;
 use tool_lala\related_data;
+use tool_lala\related_data_helper;
 
 /**
  * Class for the output for a single model version.
@@ -64,7 +65,7 @@ class evidence_item implements templatable, renderable {
         $data['timecollectionfinished'] = userdate((int) $this->item->timecollectionfinished);
         $data['serializedfilelocation'] = $this->item->serializedfilelocation;
 
-        $tablename = related_data::get_tablename_from_serializedfilelocation($this->item->serializedfilelocation);
+        $tablename = related_data_helper::get_tablename_from_serializedfilelocation($this->item->serializedfilelocation);
         if ($tablename !== false) {
             $data['tablename'] = $tablename;
         }
